requirements.txt
Django==1.4.4
argparse==1.2.1
distribute==0.6.31
dj-database-url==0.2.1
gevent==0.13.8
gevent-socketio==0.3.5-rc2
gevent-websocket==0.3.6
greenlet==0.4.0
gunicorn==0.16.1
psycopg2==2.4.6
wsgiref==0.1.2
